Separate events by years
Events do not need to be uniquely named in shotwell, so it is
possible for duplicate named events to overlap each other. We
mitigate against this (and organize our output more cleanly)
by placing events within a parent directory named for the year
the photos were taken in (using the primary image exposure time).